+++ test/pdfs/issue/issue-202-page-vs-table-header.js
@@ -1,0 +1,7 @@
+module.exports = function(doc) {
+  doc.header().text("Page Header")
+
+  const table = doc.table({widths: [null]})
+  table.header().cell("Table header")
+  table.row().cell("Content")
+}